requirements/requirements.txt
alembic==1.6.5
amqp==5.0.6
aniso8601==9.0.1
appdirs==1.4.4
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.2.2
cffi==1.15.0
chardet==4.0.0
click==8.0.0
cryptography==3.4.8
datumaro==0.2.1
defusedxml==0.7.1
distlib==0.3.2
dnspython==1.16.0
eventlet==0.30.2
filelock==3.0.12
Flask==2.0.1
Flask-JWT-Extended==3.24.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.7.0
importlib-resources==5.2.2
itsdangerous==2.0.1
kombu==5.2.2
lxml==4.6.5
Mako==1.1.5
MarkupSafe==2.0.1
monotonic==1.6
packaging==21.3
paramiko==2.7.2
pg8000==1.21.1
psutil==5.8.0
pkginfo==1.8.2
py-cpuinfo==8.0.0
pycocotools==2.0.2
pycparser==2.21
PyJWT==1.7.1
PyNaCl==1.4.0
python-editor==1.0.4
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
scramp==1.4.1
sentencepiece==0.1.96
SQLAlchemy==1.4.23
urllib3==1.26.8
vine==5.0.0
virtualenv==20.7.2
wcwidth==0.2.5
Werkzeug==2.0.1
zipp==3.5.0
huggingface_hub==0.4.0
transformers[onnx]==4.16.2
